refactor: 继续迁移 Command 模式的功能,将通用的操作移动至 KeyEventCallback. feat: f/F char 查找指定字符 impl nG/G b
fmt

fix multiple import

以单词为目标的移动

rename

1

fix: w移动越界

feat: f/F char 查找指定字符
